Issue #3 Fix
These changes ensure there are not unncessary deletes. The change is
simple (just make sure if you have both a matched and linked object that
they are actually different before you do the delete). The test is a
little less obvious.

+++ src/nsync/actions.py
@@ -301,8 +301,10 @@ class UpdateModelWithReferenceAction(UpdateModelAction):
         if self.find_objects().exists():
             matched_object=self.find_objects().get()
 
-        # If both matched and linked objects exist, get rid of the matched
-        if matched_object and linked_object:
+        # If both matched and linked objects exist but are different,
+        # get rid of the matched one
+        if matched_object and linked_object and (matched_object != 
+                                                 linked_object):
             matched_object.delete()
 
         # Choose the most appropriate object to update
